Tansaku/ml_requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
cmake==3.27.2
datasets==2.14.4
dill==0.3.7
dnspython==2.4.2
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
gdcm==1.1
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
joblib==1.3.2
lazy_loader==0.3
lit==16.0.6
loguru==0.7.0
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
# nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
# nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pinecone-client==2.2.2
pyarrow==12.0.1
pydicom==2.4.2
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.2
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.8.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
transformers==4.31.0
# triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
xxhash==3.3.0
yarl==1.9.2
